pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "VarNoiseDB"
version = "0.1.0"
description = "A tool for managing allele depth noise from sequencing of tumors."
authors = ["Adam Rosenbaum <adam.rosenbaum@regionvasterbotten.se>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "varnoisedb", from = "src" }]

include = ["default_config.yaml", "src/varnoisedb/data/mock.g.vcf"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
click = "^8.0.0"
pyyaml = "^6.0.1"
cerberus = "^1.3.4"
cyvcf2 = "^0.30"
psycopg2-binary = "^2.9.1"
mysql-connector-python = "^8.0.26"

[tool.poetry.scripts]
VarNoiseDB = "varnoisedb.cli.cli:cli"
